Add files via upload
This program takes a username and password and adds a salt and pepper to the hash value of the password. Right now it is using the SHA512 hashing algorithm, but that can be changed to anything that the hashlib module allows. There is also a lookup function that takes user input for a username and password and searches the respective CSV files for the username and hash and compares the hash values and username.