Add files via upload
In the prior version, there was no check for username correctness, so if (n) people had the same password only the first one would show up as existing. With this version, I added a username correctness check so after it checks the password hash it then checks the username to make sure that is correct. This allows multiple users to have the same password.